\name{print.UniProtKeywords_info}
\alias{print.UniProtKeywords_info}
\title{
Print the UniProtKeywords object
}
\description{
Print the UniProtKeywords object
}
\usage{
\method{print}{UniProtKeywords_info}(x, ...)
}
\arguments{

  \item{x}{A \code{UniProtKeyword} object.}
  \item{...}{Other arguments}

}
\examples{
UniProtKeywords
}
